refactor(db): split the transaction registry into a module
