+++ SURVEY.md
@@ -26,11 +26,11 @@ Sources:
 | `cc6eb7a3` | #4810 | bot: Update dependencies (bulk dependabot) | not-applicable | Rust crate updates run on their own cadence. |
 | `046e6caa` | #4809 | feat: Gate commit queries with DAC | **already-ported** | Rust commit `3ffc3e88` (Feb 23) — added per-doc ACP filtering at `crates/query/src/runner/commits.rs:710-792`. Test coverage in `tools/integration-test/tests/acp/audit.rs::cid_time_travel_acp_bypass_test`. Caveat: Go organizes commit-ACP tests in `tests/integration/acp/dac/commits/`; Rust consolidates them in `audit.rs`. Optional follow-up: split into `tests/acp/commits/` subdir for parity. |
 | `ee87a529` | #4796 | test(i): Exclude LevelDB from multi-txn test | not-applicable | Documents a Go-side LevelDB limitation (#4795). Rust storage backends (redb, fjall, rocksdb, memory) — none are LevelDB. |
-| `04a2d1eb` | #4688 | feat: Add syntax sugar `DeleteCollection` cmd | **port-needed** | Underlying `delete_collection` already exists in Rust at `crates/db/src/collection_ops/delete.rs:68`; HTTP DELETE route wired at `crates/http/src/handlers/collections.rs` + `routes.rs:81`. CLI wrapper missing. Add `Delete(CollectionDeleteArgs)` to `CollectionCommand` in `crates/cli/src/commands/client/collection/mod.rs:37`, plus HTTP client helper. |
+| `04a2d1eb` | #4688 | feat: Add syntax sugar `DeleteCollection` cmd | **port-needed** | Underlying `delete_collection` already exists in Rust at `crates/db/src/collection/ops/delete.rs:72`; HTTP DELETE route wired at `crates/http/src/handlers/collections.rs` + `routes.rs:81`. CLI wrapper missing. Add `Delete(CollectionDeleteArgs)` to `CollectionCommand` in `crates/cli/src/commands/client/collection/mod.rs:37`, plus HTTP client helper. |
 | `f550ce2b` | #4767 | bot: Update dependencies | not-applicable | Dependabot bulk update. |
 | `83b141c4` | #4681 | fix: Fix cli commands markdown output | not-applicable | Go-specific cobra markdown generation; Rust CLI (clap) doesn't use this tooling. |
 | `702aee24` | #4737 | test: Document incremental lens migration bug under indexes | **low-priority doc-only** | Documents Go bug #4736 with skipped tests. Track as future Rust issue if/when lens+index combo is implemented; no immediate port. |
-| `630f4f30` | #4763 | fix: Prevent txn-actions from executing concurrently | **already-ported** | Rust commit `b6255b00` in PR #951 ("fix: serialize actions per transaction"). Per-txn `action_lock: Arc<async_lock::Mutex<()>>` at `crates/db/src/txn_context.rs:29`, acquired in `crates/query/src/runner/executor.rs:302-306` and `crates/db/src/txn_registry.rs`. Test: `execute_in_txn_serializes_concurrent_actions_on_same_handle`. |
+| `630f4f30` | #4763 | fix: Prevent txn-actions from executing concurrently | **already-ported** | Rust commit `b6255b00` in PR #951 ("fix: serialize actions per transaction"). Per-txn `action_lock: Arc<async_lock::Mutex<()>>` at `crates/db/src/txn/context.rs:32`, acquired in `crates/query/src/runner/executor.rs:302-306` and `crates/db/src/txn/registry/`. Test: `execute_in_txn_serializes_concurrent_actions_on_same_handle`. |
 | `5d03d037` | #4775 | refactor: Use JS txn directly instead of context | not-applicable | JS-client-specific refactor; Rust has no JS client. |
 | `4ac67d81` | #4761 | fix: Make silent build tag work on Android and Linux | not-applicable | Go build-tag specific. |
 | `84b2398f` | #4765 | test: Introduce C binding smoke test | not-applicable | Out of FFI scope (Rust FFI is client-only per [feedback_ffi_scope](../../.claude/projects/-Users-johnzampolin-go-src-github-com-sourcenetwork-defradb-rs/memory/feedback_ffi_scope.md)). |
@@ -39,7 +39,7 @@ Sources:
 | `30cf74a9` | #4746 | test: Add signed-docs test multiplier | **file as sub-issue** | Rust already has focused signing coverage (`tools/integration-test/tests/encryption/block_verify.rs`, `tests/p2p_iroh/connection/signature.rs`). Go's multiplier value is regression-detection across the whole suite via `DEFRA_MULTIPLIERS=signed-docs`. Porting requires extending `defra-harness` (external repo `sourcenetwork/backbone`), so track separately. |
 | `ac820098` | #4754 | test: Migrate commit txn action to testo | not-applicable | Go test-framework migration to "testo"; Rust uses a different structure (`tools/integration-test/tests/*`). |
 | `7ca3b280` | #4749 | test: Migrate UpdateDoc action to new system | not-applicable | Same as above. |
-| `dc575858` | #4750 | fix: Seed CollectionVersions from DB | not-applicable | Go fix is for a test-harness reconstruction issue (change detector across process boundaries). Rust auto-loads collection versions during `DB::open_with_options()` via `load_collections()` in `crates/db/src/collection_ops/mod.rs:56-243`. Underlying durability is already correct. |
+| `dc575858` | #4750 | fix: Seed CollectionVersions from DB | not-applicable | Go fix is for a test-harness reconstruction issue (change detector across process boundaries). Rust auto-loads collection versions during `DB::open_with_options()` via `load_collections()` in `crates/db/src/collection/ops/mod.rs:56-259`. Underlying durability is already correct. |
 | `43250ed7` | #4723 | chore: Collection version templates in tests | not-applicable | Go test-template feature; doesn't map to Rust test layout. |
 | `83af37a9` | #4639 | fix: Reduce telemetry logs | not-applicable | Go fix attaches an `otel.SetErrorHandler` to dedupe OTLP collector "connection refused" spam. Rust has **no OTEL exporter setup** today — there are no `opentelemetry`/`otlp` crate uses in `crates/*/Cargo.toml`. Nothing in the Rust codebase emits these errors. Port becomes relevant only if/when an OTEL exporter is added in Rust. |
 | `a81825d4` | #4701 | chore: Fix the C binding Linux build command | not-applicable | Out of FFI scope. |
